feat(agent): create AppointmentSection component to display upcoming appointments
